fix the file name conventions and refactor the code
