+++ src/background-script/on-install-hooks.ts
@@ -9,7 +9,7 @@ import {
     blacklist,
 } from '../blacklist/background'
 import { trackExistingTabs } from '../activity-logger/background/tab-bridge'
-import CountlyAnalytics from '../analytics/countly'
+import countlyAnalytics from '../analytics/countly'
 
 export async function onInstall() {
     const now = Date.now()
@@ -19,7 +19,7 @@ export async function onInstall() {
 
     analytics.trackEvent({ category: 'Global', action: 'Install' }, true)
 
-    CountlyAnalytics.trackEvent({ key: 'install' })
+    countlyAnalytics.trackEvent({ key: 'install' })
 
     // Open onboarding page
     browser.tabs.create({ url: `${OVERVIEW_URL}?install=true` })
